New version

+++ regular.cpp
@@ -58,7 +58,7 @@ struct Vertex_Regular {
 };
 
 struct Edge_Regular {
-    float dist=1.0;
+    double dist=1.0;
     float edist=0.0;
     double distance=1.0;
     double ot=1.0;
